requirements.txt
# Core async
asyncio-throttle==1.0.2
aiohttp==3.9.5
websockets==12.0

# Exchange connectivity
ccxt==4.3.60
bitmex==1.0.3

# Database
sqlalchemy==2.0.30
aiosqlite==0.20.0

# Config
tomli==2.0.1
python-dotenv==1.0.1

# Dashboard
streamlit==1.35.0
pandas==2.2.2
plotly==5.22.0

# Notifications
python-telegram-bot==21.3

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-mock==3.14.0

# Utilities
structlog==24.2.0
